Application controller update

+++ app/Http/Controllers/applicationController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\application;
 use App\Helper\Token;
-
 class applicationController extends Controller
 {
     /**
@@ -17,7 +16,6 @@ class applicationController extends Controller
     {
         //
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -27,7 +25,6 @@ class applicationController extends Controller
     {
         //
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -37,15 +34,13 @@ class applicationController extends Controller
     public function store(Request $request)
     {
         $application = new application();
-        if(!$application->applicationExists($request->name))
-        {
+        if(!$application->applicationExists($request->name)){
             $application->new_application($request);
-        return response()->json(["Success" => "Se ha añadido la aplicacion correctamente"]);
+            return response()->json(["Success" => "Se ha añadido la aplicacion."]);
         }else{
-            return response()->json(["Error" => "La aplicacion no se ha añadido correctamente"]);
+            return response()->json(["Error" => "La aplicacion ya existe"]);
         }
     }
-
     /**
      * Display the specified resource.
      *
@@ -54,9 +49,15 @@ class applicationController extends Controller
      */
     public function show($id)
     {
-        //
+        $application = new application();
+        $applications = $application->getApplications();
+        if(isset($applications)){
+           
+            return response()->json(["Success" => $applications]);
+        }else{
+            return response()->json(["Error" => "No hay aplicaciones guardadas"]);
+        }
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -67,7 +68,6 @@ class applicationController extends Controller
     {
         //
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -77,22 +77,34 @@ class applicationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $application = application::where('name', $request->name)->first();
-        $application->name = $request->new_name;
-        $application->icon = $request->new_icon;
-        $application->update();
+        $application = application::where('name',$request->name)->first();
+        if (isset($application)) {
+            
+            $application->name = $request->name;
+            $application->icon = $request->icon;
+            $application->update();
+        
+            return response()->json(["Success" => "Se ha modificado la aplicacion."]);
+        }else{
+            return response()->json(["Error" => "La aplicacion no existe"]);
+        }
     }
-
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        $application = application::where('name', $request->name)->first();
         
-        $application->delete();
+        $application = application::where('name',$request->name)->first();
+        if (isset($application)) {
+            $application->delete();
+        
+            return response()->json(["Success" => "Se ha borrado la aplicacion."]);
+        }else{
+            return response()->json(["Error" => "La aplicacion no existe"]);
+        }
     }
 }
